refactor : 공통 관리 모듈로 이동
